feat(mypage): 마이페이지의 nav바 클릭시 정보 보여주기 위한 component 분리
